Refactored routes into their own files. Moved logic from events endpoints into a controller and repositories.
