+++ .gitignore
@@ -1,1 +1,2 @@
 *.code-workspace
+.vscode/
