+++ api/src/repositories/mod.rs
@@ -1,0 +1,2 @@
+pub mod event;
+pub mod invitation;
